Redundant cast.

+++ RCS.WpfShop.Common.Test/ConvertersTest/WeightFormatterTest.cs
@@ -10,7 +10,7 @@ namespace RCS.WpfShop.Common.Converters.Test
         public void ConvertTest()
         {
             var target = new WeightFormatter();
-            const decimal weight = (decimal)1;
+            const decimal weight = 1;
             const string unit = "Some Unit";
 
             var result = target.Convert(new object[] {weight, unit }, null, null, null);
